Ignore embedded-cassandra.yaml
It contains randomly generated port numbers, which cause cache misses.
Unfortunately, the GE Maven Extension does not yet allow for normalization of individual yaml properties, so ignoring the entire file is required to resolve the cache miss, even though it is not the "correct" way to deal with this issue.